Move the last functions out of rcore
